+++ api/Common/Shared.Core/Entities/Product.cs
@@ -31,7 +31,7 @@ public class Product : AggregateRoot<int>
         Enabled = true;
     }
 
-    public void Update(int? parentId, int? categoryId, decimal? price, string? name, string? description, List<(int,int)>? consumptions)
+    public void Update(int? parentId, int? categoryId, decimal? price, string? name, string? description, List<(int, decimal)>? consumptions)
     {
         Guard.Operation(Enabled == true, "A product cannot be modified when it's not active. Restore it and try again.");
         Guard.Operation(parentId == null || categoryId == null, "A product can either have a parent, or a category");
